Create Arduino switch
Edward good evening, As you can see the code that I added is the serial write (2) and serial write (1).  Your code is not working because the BYTE code doesn't support Arduino anymore. The 1 and 2 means in the serial write is when the user press the switch it print 1 else if the user doesn't click the switch it will print 2. Can you understand it Edward?  Just email me back if you understand it OK and have a goo day!